+++ resources/views/components/ui/datatable-js.blade.php
@@ -116,4 +116,5 @@
     </div>
 
     @include('mrcatz::components.ui.partials.modals')
+    @include('mrcatz::components.ui.datatable-bulk-action')
 </div>